+++ src/LiveComponent/assets/test/tools.ts
@@ -434,6 +434,7 @@ export function initComponent(props: any = {}, controllerValues: any = {}) {
         ${controllerValues.fingerprint ? `data-live-fingerprint-value="${controllerValues.fingerprint}"` : ''}
         ${controllerValues.listeners ? `data-live-listeners-value="${dataToJsonAttribute(controllerValues.listeners)}"` : ''}
         ${controllerValues.browserDispatch ? `data-live-browser-dispatch="${dataToJsonAttribute(controllerValues.browserDispatch)}"` : ''}
+        ${controllerValues.queryMapping ? `data-live-query-mapping-value="${dataToJsonAttribute(controllerValues.queryMapping)}"` : ''}
     `;
 }
 
@@ -449,4 +450,12 @@ export function getComponent(element: HTMLElement|null) {
     }
 
     return component;
+}
+
+export function setCurrentSearch(search: string){
+    history.replaceState(history.state, '', window.location.origin + window.location.pathname + search);
+}
+
+export function expectCurrentSearch (){
+    return expect(decodeURIComponent(window.location.search));
 }